fix: ssg implementation
