+++ app/(withlayout)/faculty/[id]/layout.tsx
@@ -1,0 +1,8 @@
+
+export default function FacultyLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return children;
+}
